+++ stdext/result.ml
@@ -4,6 +4,10 @@ let bind ~f r = bind r f
 
 let map ~f = map f
 
+(* some useful functions inspired by Jane Street Base *)
+
+let of_option v ~error = match v with Some v -> Ok v | None -> Error error
+
 module O = struct
   let ( >>= ) res f = bind ~f res